+++ CHANGELOG.md
@@ -102,3 +102,8 @@ All notable changes to `time-helpers` will be documented in this file
 ## 1.2.2 - 2021-07-30
 - optimize assertions methods used in `TimePeriodTest`
 - add test methods for testing all of `TimePeriods` public methods
+
+
+## 1.2.3 - 2021-07-30
+- optimize `TimePeriods::mapMethods()` & removed use of `ArrayHelpers`
+- cut sfneal/array-helpers from composer requirements